refactor: split MagicCircleCanvas for mobile readability (10k char limit)
- Extract HelpModal to src/components/HelpModal.tsx (3KB)
- Extract TutorialOverlay to src/components/TutorialOverlay.tsx (2.4KB)
- MagicCircleCanvas.tsx reduced to 11KB (was 16KB)
- Components organized by function/responsibility
