+++ src/components/MagicCircleCanvas.tsx
@@ -7,127 +7,8 @@ interface MagicCircleCanvasProps {
   onScore: (result: ScoringResult) => void;
   onReset: () => void;
 }
-
-/** ヘルプモーダル */
-function HelpModal({ onClose }: { onClose: () => void }) {
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center tutorial-overlay"
-      style={{ background: 'rgba(0,0,0,0.6)' }}
-      onClick={onClose}
-      onTouchEnd={onClose}
-    >
-      <div
-        className="mx-4 max-w-sm rounded-xl p-6"
-        style={{ background: '#1a1a2e', border: '1px solid rgba(0,229,255,0.3)' }}
-        onClick={(e) => e.stopPropagation()}
-        onTouchEnd={(e) => e.stopPropagation()}
-      >
-        <h2 className="mb-4 text-xl font-bold" style={{ color: '#00e5ff' }}>
-          🔮 ヘルプ
-        </h2>
-        <ol className="space-y-3 text-sm" style={{ color: '#c0c0e0' }}>
-          <li className="flex gap-2">
-            <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-cyan-600 font-bold text-white">1</span>
-            <span>グレーの<strong>三角形のお手本</strong>が表示されます</span>
-          </li>
-          <li className="flex gap-2">
-            <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-cyan-600 font-bold text-white">2</span>
-            <span><strong>赤い点の位置</strong>から指でなぞり始めてください</span>
-          </li>
-          <li className="flex gap-2">
-            <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-cyan-600 font-bold text-white">3</span>
-            <span>三角形の頂点に沿って<strong>一周</strong>してください（制限時間: 5秒）</span>
-          </li>
-          <li className="flex gap-2">
-            <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-cyan-600 font-bold text-white">4</span>
-            <span><strong>詠唱完了！</strong>ボタンでスコア判定されます</span>
-          </li>
-        </ol>
-        <hr className="my-4" style={{ borderColor: 'rgba(0,229,255,0.2)' }} />
-        <div className="text-xs" style={{ color: '#7676aa' }}>
-          <p className="mb-1"><strong>スコア一覧：</strong></p>
-          <p><span style={{ color: '#ffd700' }}>Sランク 120%ダメージ</span> (90点以上)</p>
-          <p><span style={{ color: '#00e5ff' }}>Aランク 100%ダメージ</span> (70点以上)</p>
-          <p><span style={{ color: '#76ff03' }}>Bランク 70%ダメージ</span> (50点以上)</p>
-          <p><span style={{ color: '#ff4081' }}>Cランク 失敗</span> (50点未満)</p>
-        </div>
-        <button
-          onClick={onClose}
-          onTouchEnd={onClose}
-          className="mt-4 w-full cursor-pointer rounded-md py-2 font-bold text-black transition-opacity hover:opacity-80"
-          style={{ background: 'linear-gradient(135deg, #00e5ff, #7c4dff)' }}
-        >
-          閉じる
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/** チュートリアルオーバーレイ（初回表示） */
-function TutorialOverlay({ onStart }: { onStart: () => void }) {
-  const [step, setStep] = useState(0);
-  const steps = [
-    {
-      title: '🔮 Arcane Tracerへようこそ！',
-      body: '魔法陣をお手本通りになぞって、正確さを競うアプリです。',
-    },
-    {
-      title: '📐 手順',
-      body: '画面に表示される灰色の三角形を、赤い点の位置から指でなぞってください。制限時間は5秒です。\n\n正確さによってS〜Cランクのスコアが出ます！',
-    },
-    {
-      title: '✨ 詠唱完了！',
-      body: '描き終わったら「詠唱完了！」ボタンを押してスコア判定を受けましょう。\n\nよい魔力を！',
-    },
-  ];
-
-  const nextStep = () => {
-    if (step < steps.length - 1) setStep(step + 1);
-    else onStart();
-  };
-
-  return (
-    <div
-      className="fixed inset-0 z-40 flex items-center justify-center tutorial-overlay"
-      style={{ background: 'rgba(0,0,0,0.85)' }}
-    >
-      <div
-        className="mx-4 max-w-sm rounded-xl p-6 text-center"
-        style={{ background: '#1a1a2e', border: '1px solid rgba(0,229,255,0.3)' }}
-      >
-        <h2 className="mb-3 text-xl font-bold" style={{ color: '#00e5ff' }}>
-          {steps[step].title}
-        </h2>
-        <p className="mb-6 whitespace-pre-line text-sm" style={{ color: '#c0c0e0' }}>
-          {steps[step].body}
-        </p>
-        <div className="mb-4 flex justify-center gap-1">
-          {steps.map((_, i) => (
-            <div
-              key={i}
-              className="h-2 w-2 rounded-full transition-all"
-              style={{ background: i <= step ? '#00e5ff' : '#333366' }}
-            />
-          ))}
-        </div>
-        <button
-          onClick={nextStep}
-          onTouchEnd={(e) => {
-            e.preventDefault();
-            e.stopPropagation();
-            nextStep();
-          }}
-          className="w-full cursor-pointer rounded-md py-2 font-bold text-black transition-opacity hover:opacity-80"
-          style={{ background: 'linear-gradient(135deg, #00e5ff, #7c4dff)' }}
-        >
-          {step < steps.length - 1 ? '次へ →' : '始める ✨'}
-        </button>
-      </div>
-    </div>
-  );
-}
+import HelpModal from './HelpModal';
+import TutorialOverlay from './TutorialOverlay';
 
 export default function MagicCircleCanvas({ onScore, onReset }: MagicCircleCanvasProps) {
   const canvasRef = useRef<HTMLCanvasElement>(null);
